requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
google-generativeai
